backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
cryptography==45.0.6
httpx==0.28.1
google-api-python-client==2.177.0
google-auth==2.40.3
openai==1.99.9
python-multipart==0.0.20
pypdf==5.9.0
python-docx==1.2.0
pytest==8.4.1
pytest-asyncio==1.1.0
